+++ packages/cli/plugin-nocode/src/register/index.ts
@@ -1,17 +1,12 @@
 import { promisify } from 'util';
 import path from 'path';
-import { fs, glob as globModule, inquirer } from '@modern-js/utils';
-import semver from 'semver';
+import { fs, glob as globModule, semver, inquirer } from '@modern-js/utils';
 import type { NormalizedConfig } from '@modern-js/core';
 import axios from './axios';
 import loggerModule from './logger';
 import { registerPackage, registerGroup, unregisterPackage } from './butter';
 import { NPM_REGISTRY } from './constants';
-import {
-  addOfficialOwner,
-  // createDependencyCache,
-  uploadUmdToTos,
-} from './butter-cache';
+import { addOfficialOwner } from './butter-cache';
 // import preValidate from './pre-validate';
 
 const logger = loggerModule.scope('REGISTER');
@@ -158,9 +153,6 @@ export const register = async (
 
   if (mode === 'register') {
     addOfficialOwner(appDirectory);
-    // TODO: 先暂时关闭
-    // createDependencyCache(appDirectory);
-    uploadUmdToTos(appDirectory);
   }
 
   try {