+++ packages/runtime/src/loader/useLoader.ts
@@ -43,7 +43,7 @@ export interface LoaderOptions<
   initialData?: TData;
 
   /**
-   * wheather skip loader
+   * whether skip loader
    * if true, the loader will not exec.
    */
   skip?: boolean;
@@ -54,7 +54,7 @@ export interface LoaderOptions<
   params?: Params;
 
   /**
-   * wheather loader can exec on build phase.
+   * whether loader can exec on build phase.
    */
   static?: boolean;
 }
@@ -90,7 +90,7 @@ const useLoader = <TData = any, Params = any, E = any>(
 
             return Promise.resolve(res);
           } catch (e) {
-            return Promise.reject(e instanceof Error ? e.message : e);
+            return Promise.reject(e);
           }
         },
         {