+++ packages/cli/core/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @modern-js/core
 
+## 1.10.1
+
+### Patch Changes
+
+- 69a728375: fix: remove exports.jsnext:source after publish
+- Updated dependencies [cd7346b0d]
+- Updated dependencies [69a728375]
+  - @modern-js/utils@1.7.2
+  - @modern-js/load-config@1.3.2
+
+## 1.10.0
+
+### Minor Changes
+
+- 0ee4bb4e: feat: prebundle webpack loaders and plugins
+
+### Patch Changes
+
+- a22d3ea8: fix: package manager name in logs and comments
+- 5c00db22: fix: ignore existed schema properties
+- 92f4909e: fix: cli.init should return correct appContext
+- Updated dependencies [0ee4bb4e]
+- Updated dependencies [6fa74d5f]
+  - @modern-js/utils@1.7.0
+
 ## 1.9.0
 
 ### Minor Changes