+++ packages/cli/plugin-router/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @modern-js/plugin-router
 
+## 1.2.13
+
+### Patch Changes
+
+- 69a728375: fix: remove exports.jsnext:source after publish
+- Updated dependencies [cd7346b0d]
+- Updated dependencies [69a728375]
+  - @modern-js/utils@1.7.2
+
+## 1.2.12
+
+### Patch Changes
+
+- a4c5fe78: fix test case
+- Updated dependencies [0ee4bb4e]
+- Updated dependencies [6fa74d5f]
+  - @modern-js/utils@1.7.0
+
 ## 1.2.11
 
 ### Patch Changes