+++ packages/cli/plugin-testing/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @modern-js/plugin-testing
 
+## 1.5.4
+
+### Patch Changes
+
+- 0e0537005: fix: unlock @babel/core version
+- 69a728375: fix: remove exports.jsnext:source after publish
+- Updated dependencies [b7b8075dc]
+- Updated dependencies [cd7346b0d]
+- Updated dependencies [0e0537005]
+- Updated dependencies [738c55d39]
+- Updated dependencies [69a728375]
+- Updated dependencies [0f86e133b]
+  - @modern-js/webpack@1.6.2
+  - @modern-js/runtime-core@1.4.5
+  - @modern-js/utils@1.7.2
+  - @modern-js/testing@1.5.1
+  - @modern-js/babel-compiler@1.2.4
+  - @modern-js/testing-plugin-bff@1.4.2
+  - @modern-js/bff-utils@1.2.5
+
+## 1.5.3
+
+### Patch Changes
+
+- 3f209269: feat(testing): add `afterTest` hook for testing plugin
+- Updated dependencies [4697d1db]
+- Updated dependencies [0ee4bb4e]
+- Updated dependencies [6fa74d5f]
+- Updated dependencies [3f209269]
+  - @modern-js/webpack@1.6.0
+  - @modern-js/testing@1.5.0
+  - @modern-js/utils@1.7.0
+  - @modern-js/runtime-core@1.4.4
+  - @modern-js/testing-plugin-bff@1.4.1
+
 ## 1.5.1
 
 ### Patch Changes