+++ packages/cli/plugin-unbundle/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @modern-js/plugin-unbundle
 
+## 1.4.1
+
+### Patch Changes
+
+- 0e0537005: fix: unlock @babel/core version
+- 69a728375: fix: remove exports.jsnext:source after publish
+- Updated dependencies [cd7346b0d]
+- Updated dependencies [0e0537005]
+- Updated dependencies [69a728375]
+  - @modern-js/utils@1.7.2
+  - @modern-js/server@1.4.13
+  - @modern-js/esmpack@1.3.1
+  - @modern-js/css-config@1.2.5
+  - @modern-js/bff-utils@1.2.5
+  - @modern-js/create-request@1.2.7
+
+## 1.4.0
+
+### Minor Changes
+
+- 0ee4bb4e: feat: prebundle webpack loaders and plugins
+
+### Patch Changes
+
+- a22d3ea8: fix: package manager name in logs and comments
+- Updated dependencies [4697d1db]
+- Updated dependencies [0ee4bb4e]
+- Updated dependencies [6fa74d5f]
+  - @modern-js/create-request@1.2.6
+  - @modern-js/esmpack@1.3.0
+  - @modern-js/utils@1.7.0
+  - @modern-js/server@1.4.12
+  - @modern-js/css-config@1.2.4
+
 ## 1.3.10
 
 ### Patch Changes