+++ packages/cli/webpack/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @modern-js/webpack
 
+## 1.6.2
+
+### Patch Changes
+
+- b7b8075dc: feat: bump html-loader from v2 to v3
+- 0e0537005: fix: unlock @babel/core version
+- 738c55d39: fix: incorrect tapable version in some cases
+- 69a728375: fix: remove exports.jsnext:source after publish
+- Updated dependencies [cd7346b0d]
+- Updated dependencies [0e0537005]
+- Updated dependencies [69a728375]
+  - @modern-js/utils@1.7.2
+  - @modern-js/babel-preset-app@1.3.1
+  - @modern-js/css-config@1.2.5
+
+## 1.6.0
+
+### Minor Changes
+
+- 0ee4bb4e: feat: prebundle webpack loaders and plugins
+
+### Patch Changes
+
+- 4697d1db: fix: remove webpack-node-external for server bundle
+- Updated dependencies [0ee4bb4e]
+- Updated dependencies [6fa74d5f]
+  - @modern-js/babel-preset-app@1.3.0
+  - @modern-js/utils@1.7.0
+  - @modern-js/css-config@1.2.4
+
 ## 1.5.7
 
 ### Patch Changes