+++ packages/review/testing/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @modern-js/testing
 
+## 1.5.1
+
+### Patch Changes
+
+- 0e0537005: fix: unlock @babel/core version
+- 69a728375: fix: remove exports.jsnext:source after publish
+- Updated dependencies [cd7346b0d]
+- Updated dependencies [0e0537005]
+- Updated dependencies [69a728375]
+  - @modern-js/utils@1.7.2
+  - @modern-js/babel-preset-app@1.3.1
+
+## 1.5.0
+
+### Minor Changes
+
+- 0ee4bb4e: feat: prebundle webpack loaders and plugins
+
+### Patch Changes
+
+- 3f209269: feat(testing): add `afterTest` hook for testing plugin
+- Updated dependencies [0ee4bb4e]
+- Updated dependencies [6fa74d5f]
+  - @modern-js/babel-preset-app@1.3.0
+  - @modern-js/utils@1.7.0
+
 ## 1.4.4
 
 ### Patch Changes