+++ packages/server/prod-server/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @modern-js/prod-server
 
+## 1.1.4
+
+### Patch Changes
+
+- 69a728375: fix: remove exports.jsnext:source after publish
+- Updated dependencies [cd7346b0d]
+- Updated dependencies [69a728375]
+  - @modern-js/utils@1.7.2
+  - @modern-js/server-core@1.3.4
+
+## 1.1.3
+
+### Patch Changes
+
+- a4c5fe78: fix test case
+- 6fa74d5f: add internal metrics and logger
+- Updated dependencies [0ee4bb4e]
+- Updated dependencies [6fa74d5f]
+  - @modern-js/utils@1.7.0
+  - @modern-js/server-core@1.3.3
+
 ## 1.1.2
 
 ### Patch Changes