+++ packages/server/server/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @modern-js/server
 
+## 1.4.13
+
+### Patch Changes
+
+- 0e0537005: fix: unlock @babel/core version
+- 69a728375: fix: remove exports.jsnext:source after publish
+- Updated dependencies [cd7346b0d]
+- Updated dependencies [0e0537005]
+- Updated dependencies [69a728375]
+  - @modern-js/utils@1.7.2
+  - @modern-js/server-utils@1.2.5
+  - @modern-js/bff-utils@1.2.5
+  - @modern-js/hmr-client@1.2.6
+  - @modern-js/prod-server@1.1.4
+
+## 1.4.12
+
+### Patch Changes
+
+- 6fa74d5f: add internal metrics and logger
+- Updated dependencies [0ee4bb4e]
+- Updated dependencies [a4c5fe78]
+- Updated dependencies [6fa74d5f]
+  - @modern-js/utils@1.7.0
+  - @modern-js/prod-server@1.1.3
+  - @modern-js/server-utils@1.2.4
+
 ## 1.4.11
 
 ### Patch Changes