+++ packages/toolkit/compiler/babel/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @modern-js/babel-compiler
 
+## 1.2.4
+
+### Patch Changes
+
+- 0e0537005: fix: unlock @babel/core version
+- 69a728375: fix: remove exports.jsnext:source after publish
+- Updated dependencies [cd7346b0d]
+- Updated dependencies [69a728375]
+  - @modern-js/utils@1.7.2
+
 ## 1.2.3
 
 ### Patch Changes